Moving towards a reusable Book Item in the list
